backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
mysql-connector-python==8.1.0
cx_Oracle==8.3.0
pyodbc==5.1.0
pandas==2.0.3
python-dotenv==1.0.0